Add SQLite migrations
The previous migrations generated by Flask-migration did not work when
using SQLite as DMS, because it does not support teh deletion of
columns. As a workaround the migration repository
backend/sqlite_migrations was setup, which also works with SQLite.
To use the new repository with SQLite the Migrate() constructor has to
be called with render_as_batch=True. backend/app/__init__.py is updated
to use this call, if SQLALCHEMY_DATABASE_URI indicates a SQLite backend.