Un-predefine used SocketIO async_mode
Previously the async_mode of SocketIO was predefined as "gevent" in the
constructor SocketIO(). This did not work for all setups, because of
bugs of/with gevent, that may be caused by wrong versions of the libary.
SocketIO now determines its async_mode by itself.
Some calls to print(), that were used for debugging, are removed.