Remove sending feedback as e-mail
Remove the feature to send feedback to the developers directly.
In the frontend the form, that allowed users to enter feedback is
removed. In the backend the routes, that took this feedback and send it
to the developers via e-mail, is removed. The configuration files are
cleaned from the now unnecessary options.